requirements.txt
faker
opentelemetry-distro
psycopg2-binary
